python/requirements.txt
requests
sseclient
pyzmq